backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
pyserial==3.5
python-socketio==5.10.0
